// ui/src/types/MaterialFlow.ts
export type MaterialFlow = {
  country: string;
  companyCode: string;
  plant: string;
  material: string;
  fiscalYear: number;
  postingPeriod: number;
  processCategory: string;
  torM: string;
  senderPlant: string;
  senderMaterial: string;
  detailKey: string;
  gvcQuantity: number;
  baseUnitOfMeasure: string;
  currency: string;
};